@@ -1,6 +1,6 @@
 {
   "name": "eyed",
-  "version": "3.1.0",
+  "version": "3.2.0",
   "main": "dist/main.cjs.js",
   "types": "dist/main.d.ts",
   "scripts": {
@@ -9,9 +9,7 @@
     "pretest": "npm run lint",
     "test": "npm run testonly",
     "testonly": "mocha ./test",
-    "lintsource": "eslint --fix ./src",
-    "linttests": "eslint --fix ./test",
-    "lint": "npm run lintsource && npm run linttests",
+    "lint": "eslint",
     "upgrade": "npx npm-check --update"
   },
   "author": {
@@ -28,27 +26,27 @@
   },
   "license": "MIT",
   "dependencies": {
-    "base-x": "^4.0.0",
+    "base-x": "^5.0.1",
     "uuid": "^10.0.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.38.0",
     "@types/chai": "^4.3.16",
     "@types/mocha": "^10.0.7",
-    "@types/node": "^20.14.9",
+    "@types/node": "^24.8.1",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^7.14.1",
-    "@typescript-eslint/parser": "^7.14.1",
     "chai": "^4.3.10",
-    "eslint": "^8.54.0",
-    "eslint-plugin-import": "^2.29.1",
+    "eslint": "^9.38.0",
+    "globals": "^16.4.0",
     "mocha": "^10.5.2",
     "source-map-loader": "^5.0.0",
     "source-map-support": "^0.5.21",
-    "ts-loader": "^9.5.1",
+    "ts-loader": "^9.5.4",
     "ts-node": "^10.9.2",
-    "typescript": "^5.5.2",
-    "webpack": "^5.92.1",
-    "webpack-cli": "^5.1.4"
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.46.1",
+    "webpack": "^5.102.1",
+    "webpack-cli": "^6.0.1"
   },
   "files": [
     "dist/"
